requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pandas
openpyxl
python-multipart
